Alteracao Bigdecimal pra double

+++ src/main/java/com/dev/hakeem/myfinanceapp/entity/Conta.java
@@ -6,10 +6,8 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-
-import java.math.BigDecimal;
 import java.util.Objects;
-import java.util.UUID;
+
 
 
 @AllArgsConstructor
@@ -21,13 +19,17 @@ public class Conta {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
     private Long id;
+
     @Column(name = "instituicao_Financeira",nullable = false,length = 50)
     private  String instituicaoFinanceira;
+
     @Column(name = "saldo_inicial",nullable = true,scale = 2)
-    private BigDecimal saldoInicial;
+    private double saldoInicial;
+
     @Enumerated(EnumType.STRING)
     @Column(name = "tipo_conta",nullable = false)
     private TipoConta tipoconta;
+
     @ManyToOne
     @JoinColumn(name = "user_id")
     private  User user;
